v6.7.2

+++ bot.js
@@ -196,29 +196,45 @@ bot.on('clickButton', async (button) => {
   // admin buttons
   // generic cancel fucntion for all admin button aided embeds.
   if (button.id == "admincancel") {
-    button.defer();
-    button.message.delete();
-    button.channel.send(`Cancelled channel setup for ${button.guild.name}.`).then(msg => msg.delete({timeout: 3000})).catch(err => console.log(err));
+    if (button.clicker.member.hasPermission("ADMINISTRATOR")) {
+      button.defer();
+      button.message.delete();
+      button.channel.send(`Cancelled channel setup for ${button.guild.name}.`).then(msg => msg.delete({timeout: 3000})).catch(err => console.log(err));
+    } else {
+      button.clicker.send(`You tried to use admin only buttons in ${button.guild.name} and we thought we would let you know that you cannot do that.`);
+    };
   };
   
   // channel setup buttons
   if (button.id == "welcome") {
-    button.defer();
-    button.message.delete();
-    // set welcome channel
-    await connection.query(`UPDATE guildConfig SET welcomeChannelName = "${button.channel.name}", welcomeChannelID ="${button.channel.id}" WHERE guildID = "${button.guild.id}"`).catch(err => console.log(err));
-    button.channel.send(`Welcome channel set to ${button.channel} in ${button.guild.name}.`).then(msg => msg.delete({timeout: 3000})).catch(err => console.log(err));
+    if (button.clicker.member.hasPermission("ADMINISTRATOR")) {
+      button.defer();
+      button.message.delete();
+      // set welcome channel
+      await connection.query(`UPDATE guildConfig SET welcomeChannelName = "${button.channel.name}", welcomeChannelID ="${button.channel.id}" WHERE guildID = "${button.guild.id}"`).catch(err => console.log(err));
+      button.channel.send(`Welcome channel set to ${button.channel} in ${button.guild.name}.`).then(msg => msg.delete({timeout: 3000})).catch(err => console.log(err));
+    } else {
+      button.clicker.send(`You tried to use admin only buttons in ${button.guild.name} and we thought we would let you know that you cannot do that.`);
+    };
   } else if (button.id == "system") {
-    button.defer();
-    button.message.delete();
-    // set system channel
-    await connection.query(`UPDATE guildConfig SET systemChannelName = "${button.channel.name}", systemChannelID ="${button.channel.id}" WHERE guildID = "${button.guild.id}"`).catch(err => console.log(err));
-    button.channel.send(`System channel set to ${button.channel} in ${button.guild.name}.`).then(msg => msg.delete({timeout: 3000})).catch(err => console.log(err));
+    if (button.clicker.member.hasPermission("ADMINISTRATOR")) {
+      button.defer();
+      button.message.delete();
+      // set system channel
+      await connection.query(`UPDATE guildConfig SET systemChannelName = "${button.channel.name}", systemChannelID ="${button.channel.id}" WHERE guildID = "${button.guild.id}"`).catch(err => console.log(err));
+      button.channel.send(`System channel set to ${button.channel} in ${button.guild.name}.`).then(msg => msg.delete({timeout: 3000})).catch(err => console.log(err));
+    } else {
+      button.clicker.send(`You tried to use admin only buttons in ${button.guild.name} and we thought we would let you know that you cannot do that.`);
+    };
   } else if (button.id == "announcement") {
-    button.defer();
-    button.message.delete();
-    // set announcement channel
-    await connection.query(`UPDATE guildConfig SET announcementChannelName = "${button.channel.name}", announcementChannelID ="${button.channel.id}" WHERE guildID = "${button.guild.id}"`).catch(err => console.log(err));
-    button.channel.send(`Announcement channel set to ${button.channel} in ${button.guild.name}.`).then(msg => msg.delete({timeout: 3000})).catch(err => console.log(err));
+    if (button.clicker.member.hasPermission("ADMINISTRATOR")) {
+      button.defer();
+      button.message.delete();
+      // set announcement channel
+      await connection.query(`UPDATE guildConfig SET announcementChannelName = "${button.channel.name}", announcementChannelID ="${button.channel.id}" WHERE guildID = "${button.guild.id}"`).catch(err => console.log(err));
+      button.channel.send(`Announcement channel set to ${button.channel} in ${button.guild.name}.`).then(msg => msg.delete({timeout: 3000})).catch(err => console.log(err));
+    } else {
+      button.clicker.send(`You tried to use admin only buttons in ${button.guild.name} and we thought we would let you know that you cannot do that.`);
+    };
   };
 });
